Backend/requirements.txt
flask==2.2.5
flask_sqlalchemy==2.5.1
flask-cors==3.0.10
numpy==1.26.0
psycopg2-binary==2.9.10
scikit-learn==1.3.2
werkzeug==2.2.3
scipy==1.11.4
SQLAlchemy<2.0
pandas==1.5.3
PyJWT==2.6.0